Changed uint8 to uint16 in saving tag lengths

+++ include/pangenome_index/tag_arrays.hpp
@@ -35,20 +35,20 @@ namespace panindexer {
         void load(std::istream& in);
         void query(size_t start, size_t end);
 
-        void serialize_run_by_run(std::ofstream& out, const std::vector<std::pair<pos_t, uint8_t>>& tag_runs);
+        void serialize_run_by_run(std::ofstream& out, const std::vector<std::pair<pos_t, uint16_t>>& tag_runs);
         void deserialize_run_by_run(const std::string& file_name);
 
         void store_blocks(std::ostream& out);
-        static std::pair<pos_t, uint8_t> load_block_at(std::istream& in, size_t &next_block_start);
+        static std::pair<pos_t, uint16_t> load_block_at(std::istream& in, size_t &next_block_start);
 
         void store_blocks_sdsl(std::string filename);
-        static std::pair<pos_t, uint8_t> decode_run(gbwt::size_type decc);
+        static std::pair<pos_t, uint16_t> decode_run(gbwt::size_type decc);
 
-        vector<pair<pos_t, uint8_t>> get_tag_runs(){
+        vector<pair<pos_t, uint16_t>> get_tag_runs(){
             return tag_runs;
         };
 
-        void compressed_serialize(std::ostream &main_out, std::ostream &encoded_starts_file, std::ostream &bwt_intervals_file, std::vector<std::pair<pos_t, uint8_t>> &tag_runs);
+        void compressed_serialize(std::ostream &main_out, std::ostream &encoded_starts_file, std::ostream &bwt_intervals_file, std::vector<std::pair<pos_t, uint16_t>> &tag_runs);
         void merge_compressed_files(const std::string filename, const std::string encoded_starts_file, const std::string bwt_intervals_file);
         void load_compressed_tags(std::istream &in);
         void query_compressed(size_t start, size_t end, size_t &number_of_runs);
@@ -70,7 +70,7 @@ namespace panindexer {
         sdsl::bit_vector::select_1_type encoded_runs_starts_select;
         sdsl::sd_vector<>::rank_1_type bwt_intervals_rank;
 
-        vector<pair<pos_t, uint8_t>> tag_runs;
+        vector<pair<pos_t, uint16_t>> tag_runs;
 
 
         // Variables for the serialization
